third_party/blink/web_tests/http/tests/devtools/tracing/timeline-style/timeline-style-recalc-all-invalidator-types-expected.txt

Tests the Timeline API instrumentation of style recalc invalidator invalidations.


Running: testClassName
[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:36}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:37}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testIdWithoutStyleChange

Running: testId
[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:52}
        changedAttribute : undefined
        changedClass : undefined
        changedId : "testElementFour"
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementFour' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:53}
        changedAttribute : undefined
        changedClass : undefined
        changedId : "testElementFive"
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementFive' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testStyleAttributeChange
[
    {
        cause : {reason: StyleSheetChange, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:60}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementFour' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: StyleSheetChange, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:61}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementFive' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testAttributeChange
[
    {
        cause : {reason: Attribute, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:68}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : "dir"
        nodeName : "DIV id='testElementFour' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Attribute, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:69}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : "dir"
        nodeName : "DIV id='testElementFive' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:68}
        changedAttribute : "dir"
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementFour' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:69}
        changedAttribute : "dir"
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementFive' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testPseudoChange
[
    {
        cause : {reason: PseudoClass, stackTrace: test://evaluations/0/timeline-style-recalc-all-invalidator-types.js:78}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ":hover"
        nodeName : "DIV id='testElementFive' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
